+++ tests/Specification/ToArrayTest.php
@@ -49,13 +49,8 @@ class ToArrayTest extends TestCase
     {
         $this
             ->forAll(Set\Elements::of(
-                Sign::inequality,
                 Sign::lessThan,
                 Sign::moreThan,
-                Sign::lessThanOrEqual,
-                Sign::moreThanOrEqual,
-                Sign::isNull,
-                Sign::isNotNull,
                 Sign::startsWith,
                 Sign::endsWith,
                 Sign::contains,